feat: Split core and puppeteer features
